+++ config/application.rb
@@ -19,12 +19,9 @@ module Analytica
     config.i18n.load_path +=  Dir[Rails.root.join('config', 'locales', '*.yml')]
     config.i18n.available_locales = %i[en]
     config.i18n.default_locale = :en
-
-    # Configuration for the application, engines, and railties goes here.
-    # These settings can be overridden in specific environments using the files
-    # in config/environments, which are processed later.
-    #
     config.time_zone = "Harare"
-    # config.eager_load_paths << Rails.root.join("extras")
+    config.generators do |g|
+      g.orm :active_record, primary_key_type: :uuid
+    end
   end
 end