+++ db/migrate/20230808064933_create_active_storage_tables.active_storage.rb
@@ -51,7 +51,7 @@ class CreateActiveStorageTables < ActiveRecord::Migration[5.2]
       config = Rails.configuration.generators
       setting = config.options[config.orm][:primary_key_type]
       primary_key_type = setting || :primary_key
-      foreign_key_type = setting || :bigint
+      foreign_key_type = setting || :uuid
       [primary_key_type, foreign_key_type]
     end
 end