+++ db/migrate/20230810055932_create_lots.rb
@@ -1,13 +1,13 @@
 class CreateLots < ActiveRecord::Migration[7.0]
   def change
-    create_table :lots do |t|
+    create_table :lots, id: :uuid do |t|
       t.string :title
       t.integer :quantity
       t.integer :asking_price
       t.string :location
       t.string :condition
-      t.references :company, null: false, foreign_key: true
-      t.references :auction, null: false, foreign_key: true
+      t.references :company, null: false, foreign_key: true, type: :uuid
+      t.references :auction, null: false, foreign_key: true, type: :uuid
 
       t.timestamps
     end