+++ db/migrate/20230810062232_create_bids.rb
@@ -1,11 +1,11 @@
 class CreateBids < ActiveRecord::Migration[7.0]
   def change
-    create_table :bids do |t|
+    create_table :bids, id: :uuid do |t|
       t.integer :amount
       t.text :delivery_options
       t.string :location
-      t.references :lot, null: false, foreign_key: true
-      t.references :company, null: false, foreign_key: true
+      t.references :lot, null: false, foreign_key: true, type: :uuid
+      t.references :company, null: false, foreign_key: true, type: :uuid
 
       t.timestamps
     end