+++ db/migrate/20231211102432_devise_create_administrators.rb
@@ -2,7 +2,7 @@
 
 class DeviseCreateAdministrators < ActiveRecord::Migration[7.0]
   def change
-    create_table :administrators do |t|
+    create_table :administrators, type: :uuid do |t|
       ## Database authenticatable
       t.string :email,              null: false, default: ""
       t.string :encrypted_password, null: false, default: ""